loader: Move Windows logic to dedicated file
Relocated most windows specific functionality into loader_windows.h/.c
This removes the need for loader.c to include windows specific header files.
